sdk/pyproject.toml
[tool.poetry]
name = "sdk"
version = "0.1.0"
description = ""
authors = ["Kevin Zhang <me@kevz.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
solders = "0.21.0"
solana = "^0.35.0"
jsonref = "^1.1.0"
openai = "^1.53.0"
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
